packages/admin/pyproject.toml
[project]
name = "lineo-admin"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "J. David Ibáñez", email = "jdavid.ibp@gmail.com" }
]
license = { text = "GPL-3.0-only" }
requires-python = ">=3.12"
dependencies = [
    "django-bootstrap5",
    "django-formset",
    "django-htmx",
    "django-vite",
    "fontawesomefree"
]
classifiers = [
    "license :: osi approved :: gnu general public license v3 (gplv3)",
]

[build-system]
requires = ["uv_build>=0.8.7,<0.9.0"]
build-backend = "uv_build"
